systemd: Wait 10 seconds before restarting ceph-mgr
We do this for the MON and OSD as well, wait for a few
seconds before we try to attempt a restart.

On boot in IPv6 networks it might take a few seconds longer
before a IP-address is usable and this does not allow the mgr
to start right away.

Fixes: http://tracker.ceph.com/issues/23083

Signed-off-by: Wido den Hollander <wido@42on.com>
(cherry picked from commit c628cd083e48b6a9c61c25c0192b572ed76e9557)